% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/083-blck.R
\name{blck}
\alias{blck}
\alias{blck_data}
\title{Rolling operations}
\usage{
blck(x, fn, n = 1L, fill = NA, align = "left", ...)

blck_data(x, n = 1L)
}
\arguments{
\item{x}{\verb{[univariate vector]}

Univariate vector, numeric or ts object with only one dimension.}

\item{fn}{\verb{[function]}}

\item{n}{\verb{[positive integer(1):1L]}

Window size.}

\item{fill}{\verb{[numeric or function: NA]}

Numeric value(s) or function used to fill observations.}

\item{align}{\verb{[character(1): "left"]}

Specifying whether the index of the result should be left- or right-aligned
or centered (default) compared to the rolling window of observations.}

\item{...}{Additional arguments passed to the function \code{fn}.}
}
\value{
\itemize{
\item \code{roll()} returns a vector with the same class and attributes as the input vector.
\item \code{roll_data()} Returns a list of length length(x)/step.
}
}
\description{
Apply rolling operations over a moving window for size \code{n} and increment \code{step}.
}
\examples{

x <- seq(10, 1, -1)
}
