% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/083-blck.R
\name{blck_idx}
\alias{blck_idx}
\title{Functions used to calculate non-overlapping blocks}
\usage{
blck_idx(x, n, align)
}
\arguments{
\item{x}{data argument}

\item{n}{size of the block}

\item{align}{align}
}
\description{
Functions used to calculate non-overlapping blocks
}
