% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/083-blck.R
\name{blck_mean}
\alias{blck_mean}
\alias{blck_median}
\alias{blck_modex}
\alias{blck_sd}
\title{Non-Overapping Block Moment Calculations}
\usage{
blck_mean(x, n = 2L, fill = NA)

blck_median(x, n = 2L)

blck_modex(x, n = 2L)

blck_sd(x, n = 2L)
}
\arguments{
\item{x}{numeric vector}

\item{n}{block size}

\item{fill}{\verb{[numeric or function: NA]}

Numeric value(s) or function used to fill observations.}
}
\description{
Non-Overapping Block Moment Calculations
}
