% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{blck_sample}
\alias{blck_sample}
\title{Block}
\usage{
blck_sample(x, n = 2L, drop = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{n}{block size}

\item{drop}{reduce to vector}
}
\description{
Block
}
