% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-cycle-trends.R
\name{decompose_bn}
\alias{decompose_bn}
\title{Beveridge-Nelson decomposition}
\usage{
decompose_bn()
}
\arguments{
\item{data}{A vector of first-order integrated numeric values}

\item{nlag}{A numberic scalar for the lag-order of the autoregressive (cyclical) part}
}
\value{
Values for the stochastic trend and the stationary cycle
}
\description{
Beveridge-Nelson decomposition
}
\examples{

}
