% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-basic.R
\name{diffx-rdiffx-ldiffx}
\alias{diffx-rdiffx-ldiffx}
\alias{diffx}
\alias{rdiff}
\alias{ldiffx}
\title{Compute lagged differnces}
\usage{
diffx(x, n = 1L, order = 1L, fill = NA, fill_fun = NULL, ...)

rdiff(x, n = 1L, order = 1L, fill = NA, fill_fun = NULL, ...)

ldiffx(x, n = 1L, order = 1L, default = NA, ...)
}
\arguments{
\item{x}{\verb{[univariate vector]}

Univariate vector, numeric or ts object with only one dimension.}

\item{n}{\verb{[positive integer(1): 1L]}

Value indicating which lag to use.}

\item{order}{\verb{[positive integer(1): 1L]}

Value indicating the order of the difference.}

\item{fill}{\verb{[numeric: NA]}

Value used to fill observations.}

\item{fill_fun}{\verb{[function: NULL]}

Function used to fill observations. Uses the remaining observations as the
body of the function.}

\item{...}{Additional arguments passed on to the \code{fill_fun} function.}
}
\description{
Compute lagged differnces
}
