% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-trends.R
\name{dtrend}
\alias{dtrend}
\alias{dtrend_lin}
\alias{dtrend_quad}
\alias{dtrend_poly}
\title{Deterministic Trends}
\usage{
dtrend_lin(x, bp = NULL)

dtrend_quad(x, bp = NULL)

dtrend_poly(x, degree, bp = NULL, raw = FALSE)
}
\arguments{
\item{bp}{\verb{[positive integer(1)]}

Break points to define piecewise segments of the data.}

\item{degree}{\verb{[positive integer(1)]}

Value indicating the degree of polynomial}
}
\description{
Remove global trend information from the series.
\itemize{
\item \code{dt_lin} removes the linear trend.
\item \code{dt_quad} removes the quadratic trend.
\item \code{dt_poly} removes the nth-degree polynomial trend.
}
}
\examples{

set.seed(123)
vec2 <- cumsum(c(rnorm(20), 1+rnorm(20)))
plot.ts(vec2)

plot.ts(cbind(dtrend_lin(vec2), dtrend_lin(vec2, bp = 20)))
dtrend_quad(vec2)

}
