% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill-functions.R
\name{fill_linear}
\alias{fill_linear}
\title{Fill with "linear approximation"}
\usage{
fill_linear(body, idx, ...)
}
\arguments{
\item{body}{\verb{[numeric vector]}

The body of the vector.}

\item{idx}{\verb{[integer vector]}

the index to replace with.}

\item{...}{Further arguments passed to \verb{\link[stats]\{approx\}}}
}
\value{
Returns a vector with the same class and attributes as the input vector.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#maturing'><img src='figures/lifecycle-maturing.svg' alt='Maturing lifecycle'></a>}}{\strong{Maturing}}
}
\examples{
x <- c(5,3,2,2,5)
xlen <- length(x)
n <- 2
n <- pmin(n, xlen)
idx <- 1:n
body <- x[seq_len(xlen - n)]
fill_linear(body, idx)

}
