% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.R
\name{fill_locf}
\alias{fill_locf}
\title{Fill with "Last Observation Carried Forward"}
\usage{
fill_locf(body, idx, fail = NA)
}
\arguments{
\item{body}{\verb{[numeric vector]}

The body of the vector.}

\item{idx}{\verb{[integer vector]}

the index to replace with.}

\item{fail}{\verb{[numeric(1) or numeric vector: fill]}

In case it fails to fill some values.}
}
\value{
Returns a vector with the same class and attributes as the input vector.
}
\description{
Fill with "Last Observation Carried Forward"
}
\examples{
x <- c(5,3,2,2,5)
lagx(x, n = 2, fill_fun = fill_locf)
leadx(x, n = 2, fill_fun = fill_locf)

lagx(x, n = 2, fill_fun = fill_nocb)
leadx(x, n = 2, fill_fun = fill_nocb)


xlen <- length(x)
n <- 2
n <- pmin(n, xlen)
idx <- 1:n
body <- x[seq_len(xlen - n)]
fill_locf(body, idx, NA)


xlen <- length(x)
n <- 2
n <- pmin(n, xlen)
body <- x[-seq_len(n)]
idx <- (xlen - n + 1):xlen
fill_locf(body, idx, NA)

}
