% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.R
\name{fill_vec}
\alias{fill_vec}
\title{Replaces}
\usage{
fill_vec(body, idx, vec)
}
\description{
Replaces
}
\examples{

lagx(c(1:5), fill = ~ fill_vec(.x, .y, 1:5))

lagx(c(1:5), fill = ~ fill_vec(.x, .y, roll_mean(.x)))
}
