% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill-functions.R
\name{fill_vec_}
\alias{fill_vec_}
\title{Replaces}
\usage{
fill_vec_(body, idx, vec)
}
\description{
Replaces
}
\examples{

lagx(c(1:5), fill = ~ fill_vec(.x, .y, 1:5))

lagx(c(1:5), fill = ~ fill_vec(.x, .y, roll_mean(.x)))
}
