% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill-functions.R
\name{fill_window}
\alias{fill_window}
\title{Fill window functions}
\usage{
fill_window(fn, ...)
}
\arguments{
\item{fn}{\verb{[function]}

Window function, usually of the \code{roll}, \code{rec} or \code{blck} families.}

\item{...}{Further arguments passed to \code{fn}.}
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
}
\examples{
\dontrun{
lagx(c(1:5), fill = fill_window(rec_mean))
}
}
