% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-filter.R
\name{filter_bk}
\alias{filter_bk}
\title{Baxter-King Filter}
\usage{
filter_bk(x, fill = NA, ...)
}
\arguments{
\item{x}{\verb{[univariate vector]}

Univariate vector, numeric or ts object with only one dimension.}

\item{fill}{\verb{[numeric: NA]}

Value used to fill observations.}

\item{...}{Further arguments passed to \code{\link[mFilter]{mFilter::bkfilter}}}

\item{fill_fun}{\verb{[function: NULL]}

Function used to fill observations. Uses the remaining observations as the
body of the function.}
}
\description{
This function computes the cyclical component of the Baxter-King filter.
}
\examples{

x <- cumsum(rnorm(100))
filter_bk(x)
}
