% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-filter.R
\name{filter_boosted_hp}
\alias{filter_boosted_hp}
\title{Boosted HP filter}
\source{
This function has been retrieved and rewritten from
https://github.com/zhentaoshi/Boosted_HP_filter/blob/master/R/BoostedHP.R
}
\usage{
filter_boosted_hp(
  x,
  lambda = 1600,
  iter = TRUE,
  stopping = "nonstop",
  sig_p = 0.05,
  max_iter = 100
)
}
\arguments{
\item{x}{\verb{[univariate vector]}

Univariate vector, numeric or ts object with only one dimension.}

\item{lambda}{\verb{[numeric(1): 1600]}

Smoothness penalty parameter.}

\item{stopping:}{\verb{[character: "nonstop"]}
\itemize{
\item If stopping = "adf" or "BIC", used stopping criteria.
\item If stopping = "nonstop", iterated until max_iter
}}

\item{sig_p:}{\verb{[numeric(1): 0.05]}

The significance level of the ADF test as the stopping criterion.
It is used only when stopping == "adf".}

\item{max_iter:}{\verb{[numeric(1): 100]}

the maximum number of iterations.}
}
\value{


Returns a vector with the same class and attributes as the input vector.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
}
\examples{
unemp <- ggplot2::economics$unemploy
unemp_cycle <- filter_boosted_hp(unemp)
plotx(cbind(unemp, unemp_cycle))
}
\references{
Phillips, P.C.B. and Shi, Z. (2021), BOOSTING: WHY YOU CAN USE THE HP FILTER.
International Economic Review. https://doi.org/10.1111/iere.12495
}
