% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-filter.R
\name{filter_bw}
\alias{filter_bw}
\title{Butterworth  Filter}
\usage{
filter_bw(x, ...)
}
\arguments{
\item{x}{\verb{[univariate vector]}

Univariate vector, numeric or ts object with only one dimension.}

\item{...}{Further arguments passed to \code{\link[mFilter]{bwfilter}}.}
}
\description{
This function computes the cyclical component of the Butterworth  filter.
}
\examples{
\donttest{
unemp_cycle <- filter_bw(unemp)
plotx(unemp, unemp_cycle)
}
}
