% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-cycle-trends.R
\name{filter_hamilton}
\alias{filter_hamilton}
\title{Hamilton Filter}
\usage{
filter_hamilton(x, n = 4, horizon = 8, fill = NA, fill_fun = NULL)
}
\arguments{
\item{x}{\verb{[univariate vector]}

Univariate vector, numeric or ts object with only one dimension.}

\item{n}{\verb{[numeric(1): NULL]}}

\item{horizon}{\verb{[numeric(1): NULL]}}

\item{nfix}{\verb{[logical(1): FALSE]}}
}
\value{
Returns a vector with the same class and attributes as the input vector.
}
\description{
This function computes the cyclical component of the Hodrick-Prescot filter.
}
