% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-cycle-trends.R
\name{filter_hp}
\alias{filter_hp}
\title{Hodrick Prescot Filter}
\usage{
filter_hp(x, lambda = NULL)
}
\arguments{
\item{x}{\verb{[univariate vector]}

Univariate vector, numeric or ts object with only one dimension.}

\item{lambda}{\verb{[positive numeric(1): 1600]}

Smoothing paramter.}
}
\value{
Returns a vector with the same class and attributes as the input vector.
}
\description{
This function computes the cyclical component of the Hodrick-Prescot filter.
}
