% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-filter.R
\name{filter_tr}
\alias{filter_tr}
\title{Trigonometric regression Filter}
\usage{
filter_tr(x, ...)
}
\arguments{
\item{x}{\verb{[univariate vector]}

Univariate vector, numeric or ts object with only one dimension.}

\item{...}{Further arguments passed to \code{\link[mFilter]{mFilter::trfilter}}}
}
\description{
This function computes the cyclical component of the trigonometric regression filter.
}
\examples{
x <- cumsum(rnorm(100))
filter_tr(x)

}
