% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-stats.R
\name{gmean}
\alias{gmean}
\title{Geometric Mean value}
\usage{
gmean(x, na.rm = getOption("transx.na.rm"))
}
\arguments{
\item{x}{\verb{[univariate vector]}

Univariate vector, numeric or ts object with only one dimension.}

\item{na.rm}{\verb{[logical(1): getOption("transx.na.rm")]}

A value indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
Returns a vector with the same class and attributes as the input vector.
}
\description{
Compute the sample geometric mean.
}
