% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{lead-lag}
\alias{lead-lag}
\alias{lagx}
\alias{leadx}
\title{Compute lagged or leading values}
\usage{
lagx(x, n = 1L, default = NA, ...)

leadx(x, n = 1L, default = NA, ...)
}
\arguments{
\item{x}{Vector of values}

\item{n}{Positive integer of length 1, giving the number of positions to
lead or lag by}

\item{default}{Value used for non-existent rows. Defaults to `NA`.}
}
\description{
Find the "previous" (`lag()`) or "next" (`lead()`) values in a vector.
Useful for comparing values behind of or ahead of the current values.
This functions has been taken as they are from `dply` package, but to reduce
dependecies they are not imported, but copied instead.
}
