% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-basic.R
\name{leadx-lagx}
\alias{leadx-lagx}
\alias{lagx}
\alias{leadx}
\title{Compute lagged or leading values}
\usage{
lagx(x, n = 1L, fill = NA, fill_fun = NULL, ...)

leadx(x, n = 1L, fill = NA, fill_fun = NULL, ...)
}
\arguments{
\item{x}{\verb{[univariate vector]}

Univariate vector, numeric or ts object with only one dimension.}

\item{n}{\verb{[positive integer(1): 1L]}

Value indicating the number of positions to lead or lag by.}

\item{fill}{\verb{[numeric: NA]}

Value used to fill observations.}

\item{fill_fun}{\verb{[function: NULL]}

Function used to fill observations. Uses the remaining observations as the
body of the function.}

\item{...}{Additional arguments passed on to the \code{fill_fun} function.}
}
\value{
Returns a vector with the same class and attributes as the input vector.
}
\description{
\Sexpr[results=rd, stage=render]{rlang:::lifecycle("stable")}

Find the "previous" (\code{lagx()}) or "next" (\code{leadx()}) values in a vector.
Useful for comparing values behind of or ahead of the current values.
}
\details{
This functions has been taken and modified from the \code{dplyr} package,
however, to reduce dependecies they are not imported.
}
\examples{

vec <- c(5,3,2,2,5)
lagx(vec)
lagx(vec, fill_fun = fill_nocb)

leadx(vec)
leadx(vec, fill_fun = fill_locf)
}
