% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-stats.R
\name{modex}
\alias{modex}
\alias{modex_int}
\title{Mode value}
\usage{
modex(x, na.rm = getOption("transx.na.rm"))

modex_int(x, na.rm = getOption("transx.na.rm"))
}
\arguments{
\item{x}{\verb{[univariate vector]}

Univariate vector, numeric or ts object with only one dimension.}

\item{na.rm}{\verb{[logical(1): getOption("transx.na.rm")]}

A value indicating whether NA values should be stripped before the computation proceeds.}
}
\description{
Compute the sample median.
}
