% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-basic.R
\name{norm}
\alias{norm}
\alias{norm_minmax}
\alias{norm_logistic}
\title{Normalise}
\usage{
norm_minmax(x, na.rm = TRUE)

norm_logistic(x)
}
\arguments{
\item{x}{\verb{[univariate vector]}

Univariate vector, numeric or ts object with only one dimension.}

\item{na.rm}{\verb{[logical(1): getOption("transx.na.rm")]}

A value indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
Returns a vector with the same class and attributes as the input vector.
}
\description{
Normalise
}
