% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-outliers.R
\name{out_iqr}
\alias{out_iqr}
\title{Detect outliers with Tukey's method}
\usage{
out_iqr(x, cutoff = 1.5, fill = NA, fill_fun = NULL, ...)
}
\arguments{
\item{cutoff}{\verb{[numeric(1): 121]}}

\item{...}{further arguments passed to \code{quantile}.}
}
\description{
Detect outliers with Tukey's method
}
