% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-outliers.R
\name{out_pt}
\alias{out_pt}
\title{Detect outliers with Percentiles}
\usage{
out_pt(x, pt = c(0.1, 0.9), fill = NA, fill_fun = NULL, ...)
}
\arguments{
\item{x}{mpla}

\item{pt}{percentile}
}
\description{
Detect outliers with Percentiles
}
