% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-outliers.R
\name{out_score_z}
\alias{out_score_z}
\title{Detect outliers with zscore}
\usage{
out_score_z(x, cutoff = 3, fill = NA, fill_fun = NULL, ...)
}
\arguments{
\item{x}{\verb{[univariate vector]}

Univariate vector, numeric or ts object with only one dimension.}

\item{cutoff}{\verb{[numeric(1): 3]}}

\item{fill}{\verb{[numeric: NA]}

Value used to fill observations.}

\item{fill_fun}{\verb{[function: NULL]}

Function used to fill observations. Uses the remaining observations as the
body of the function.}

\item{...}{Additional arguments passed on to the \code{fill_fun} function.}
}
\description{
Detect outliers with zscore
}
