% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-outliers.R
\name{out_score_zrob}
\alias{out_score_zrob}
\title{Detect outliers Iglewicz and Hoaglin (1993) robust z-score method}
\usage{
out_score_zrob(x, cutoff = 3.5, fill = NA, ...)
}
\arguments{
\item{x}{\verb{[univariate vector]}

Univariate vector, numeric or ts object with only one dimension.}

\item{cutoff}{\verb{[numeric(1): 3.5]}}

\item{fill}{\verb{[numeric or function: NA]}

Value(s) of function used to fill observations.}

\item{...}{further arguments passed to \code{score}.}
}
\description{
Detect outliers Iglewicz and Hoaglin (1993) robust z-score method
}
