% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-outliers.R
\name{out_threshold}
\alias{out_threshold}
\title{Detect outliers with upper and lower threshold}
\usage{
out_threshold(x, tlow = NULL, thigh = NULL, fill = NA, fill_fun = NULL)
}
\arguments{
\item{x}{mpla}
}
\description{
Detect outliers with upper and lower threshold
}
\examples{
x <- c(1, 3, -1, 5, 10, 100)
out_trim(x, tlow = 0, fill = 0)
out_trim(x, thigh = 9, fill_fun = function(x) quantile(x, 0.9))

}
