% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{out_trim}
\alias{out_trim}
\title{Trim with Percentils}
\usage{
out_trim(x, default = NA, threshold_low = NULL, threshold_high = NULL, ...)
}
\arguments{
\item{x}{mpla}
}
\description{
Trim with Percentils
}
