% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{out_trim_pt}
\alias{out_trim_pt}
\title{Trim with Percentils}
\usage{
out_trim_pt(x, default = NA, pt = c(0.1, 0.9))
}
\arguments{
\item{x}{mpla}

\item{pt}{percentile}
}
\description{
Trim with Percentils
}
