% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-outliers.R
\name{out_zscore}
\alias{out_zscore}
\title{Detect outliers with zscore}
\usage{
out_zscore(x, cutoff = 3, fill = NA, fill_fun = NULL, ...)
}
\arguments{
\item{cutoff}{\verb{[numeric(1): 3]}}

\item{...}{further arguments passed to \code{score}.}
}
\description{
Detect outliers with zscore
}
