% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-outliers.R
\name{out_zscore_rob}
\alias{out_zscore_rob}
\title{Detect outliers Iglewicz and Hoaglin (1993) robust z-score method}
\usage{
out_zscore_rob(x, cutoff = 3.5, fill = NA, fill_fun = NULL, ...)
}
\arguments{
\item{cutoff}{\verb{[numeric(1): 3.5]}}

\item{...}{further arguments passed to \code{score}.}
}
\description{
Detect outliers Iglewicz and Hoaglin (1993) robust z-score method
}
