% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotx}
\alias{plotx}
\title{Plotting wrapper around plot.default}
\usage{
plotx(x, ...)
}
\arguments{
\item{x}{\verb{[univariate vector]}

Univariate vector, numeric or ts object with only one dimension.}

\item{...}{Further arguments used in \code{plot}.}
}
\description{
Helper function to only plot x as a line plot.
}
\keyword{internal}
