% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-root-power.R
\name{pow}
\alias{pow}
\title{Raise in the power}
\usage{
pow(x, pow, modulus = FALSE)
}
\arguments{
\item{x}{\verb{[univariate vector]}

Univariate vector, numeric or ts object with only one dimension.}

\item{pow}{power to raise to}

\item{modulus}{positive}
}
\value{
Returns a vector with the same class and attributes as the input vector.
}
\description{
Raise in the power
}
