% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-root-power.R
\name{pow_manly}
\alias{pow_manly}
\title{Manly(1971)}
\usage{
pow_manly(x, lam)
}
\arguments{
\item{x}{\verb{[univariate vector]}

Univariate vector, numeric or ts object with only one dimension.}

\item{lam}{lambda}
}
\value{
Returns a vector with the same class and attributes as the input vector.
}
\description{
The transformation was reported to be successful in transform
unimodal skewed distribution into normal distribution, but is
not quite useful for bimodal or U-shaped distribution.
}
