% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-root-power.R
\name{pow_tukey}
\alias{pow_tukey}
\title{Tukey Transformations}
\usage{
pow_tukey(x, lambda = NULL, ...)
}
\arguments{
\item{x}{\verb{[univariate vector]}

Univariate vector, numeric or ts object with only one dimension.}

\item{lambda}{\verb{[numeric(1): NULL]}

Transformation exponent, \eqn{\lambda}.}

\item{...}{Further arguments passed to \code{pow}.}
}
\value{
Returns a vector with the same class and attributes as the input vector.
}
\description{
Tukey Transformations
}
\examples{
set.seed(123)
x <- rnorm(100)

}
