% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-power-transform.R
\name{power_boxcox}
\alias{power_boxcox}
\title{#' Some title
#'
#' @param x some param
#'
#' @export
power_tukey <- function(x, lambda) {
y^lam
}
Some title}
\usage{
power_boxcox(x, lam1, lam2 = NULL)
}
\arguments{
\item{x}{some param}
}
\description{
#' Some title
#'
#' @param x some param
#'
#' @export
power_tukey <- function(x, lambda) {
y^lam
}
Some title
}
