% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-basic.R
\name{rebase}
\alias{rebase}
\alias{rebase_origin}
\title{Change the base year of an index}
\usage{
rebase(x, n = NULL)

rebase_origin(x)
}
\value{
Returns a vector with the same class and attributes as the input vector.
}
\description{
descriptio goes here
}
