% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-root-power.R
\name{root_cubic}
\alias{root_cubic}
\title{Cubic root transformation}
\usage{
root_cubic(x, ...)
}
\arguments{
\item{x}{\verb{[univariate vector]}

Univariate vector, numeric or ts object with only one dimension.}

\item{...}{Further arguments passed to \code{root}.}
}
\description{
Cubic root transformation
}
