% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-root-power.R
\name{root_sq}
\alias{root_sq}
\title{Square root transformation}
\usage{
root_sq(x, ...)
}
\arguments{
\item{x}{\verb{[univariate vector]}

Univariate vector, numeric or ts object with only one dimension.}

\item{...}{Further arguments passed to \code{root}.}
}
\description{
Square root transformation
}
