% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-basic.R
\name{score}
\alias{score}
\alias{score_z}
\alias{score_mad}
\alias{score_t}
\alias{score_chisq}
\title{Score transformation}
\usage{
score_z(x, na.rm = getOption("transx.na.rm"))

score_mad(x, na.rm = getOption("transx.na.rm"))

score_t(x, na.rm = getOption("transx.na.rm"))

score_chisq(na.rm = getOption("transx.na.rm"))
}
\value{
Returns a vector with the same class and attributes as the input vector.
}
\description{
Score transformation
}
\seealso{
outlier::scores
}
