% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-smooth.R
\name{smooth_kernel}
\alias{smooth_kernel}
\title{Kernel Regression Smoother}
\usage{
smooth_kernel(x, ...)
}
\arguments{
\item{x}{\verb{[univariate vector]}

Univariate vector, numeric or ts object with only one dimension.}

\item{...}{Further arguments passed to \code{\link[ksmooth]{smooth_kernel}}}
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
}
\examples{

x <- co2
plotx(smooth_kernel(x))
}
