% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-smooth.R
\name{smooth_ma}
\alias{smooth_ma}
\title{Moving-average smoothing}
\usage{
smooth_ma(x, order = NULL, centre = TRUE, fill = NA)
}
\arguments{
\item{x}{\verb{[univariate vector]}

Univariate vector, numeric or ts object with only one dimension.}

\item{order}{\verb{[integer(1): NULL]}

Order of moving average smoother.}

\item{centre}{\verb{[logical(1): TRUE]}

Centers the moving average for even orders.}

\item{fill}{\verb{[numeric or function: NA]}

Numeric value(s) or function used to fill observations.}
}
\value{
Returns a vector with the same class and attributes as the input vector.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

Computes a simple moving average smoother.
}
\examples{

x <- co2
x <- c(1:3, 5, 4, 7:3, 2*(2:5), rep(10, 4))#sin(1:100)
plotx(x)
lines(smooth_ma(x, 4), col = "red")
lines(smooth_spline(x), col = "purple")
lines(smooth_loess(x), col = "green")
}
