% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-smooth.R
\name{smooth_spline}
\alias{smooth_spline}
\title{Fit a Smoothing Spline}
\usage{
smooth_spline(x, ...)
}
\arguments{
\item{x}{\verb{[univariate vector]}

Univariate vector, numeric or ts object with only one dimension.}

\item{...}{Further arguments passed to \code{\link[stats:smooth.spline]{smooth.spline}}}
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
}
\examples{

x <- co2
plotx(smooth_spline(x))
}
