% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-basic.R
\name{std}
\alias{std}
\alias{std_mean}
\alias{std_median}
\title{Standarize}
\usage{
std_mean(x, na.rm = TRUE)

std_median(x, na.rm = TRUE)
}
\arguments{
\item{x}{\verb{[univariate vector]}

Univariate vector, numeric or ts object with only one dimension.}

\item{na.rm}{\verb{[logical(1): getOption("transx.na.rm")]}

A value indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
Returns a vector with the same class and attributes as the input vector.
}
\description{
Standarize
}
