% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{z_score_mod}
\alias{z_score_mod}
\title{Iglewicz and Hoaglin (1993) robust z-score method}
\usage{
z_score_mod(x, default = NA, cutoff = 3.5, ...)
}
\description{
Iglewicz and Hoaglin (1993) robust z-score method
}
