% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-basic.R
\name{demean-demedian}
\alias{demean-demedian}
\alias{demean}
\alias{demedian}
\alias{demode}
\title{Removes measure of centrality from the series}
\usage{
demean(x, na.rm = getOption("transx.na.rm"))

demedian(x, na.rm = getOption("transx.na.rm"))

demode(x, na.rm = getOption("transx.na.rm"))
}
\arguments{
\item{x}{\verb{[univariate vector]}

Univariate vector, numeric or ts object with only one dimension.}

\item{na.rm}{\verb{[logical(1): getOption("transx.na.rm")]}

A value indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
Returns a vector with the same class and attributes as the input vector.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#maturing'><img src='figures/lifecycle-maturing.svg' alt='Maturing lifecycle'></a>}}{\strong{Maturing}}

Removes the mean, the median or the mode from the series.
}
\examples{

x <- c(2,5,10,20,30)
summary(x)

demean(x)
demedian(x)
demode(x)
}
