% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{diffx-rdiffx-ldiffx}
\alias{diffx-rdiffx-ldiffx}
\alias{diffx}
\alias{rdiffx}
\alias{ldiffx}
\title{Lagged Differnces}
\usage{
diffx(x, n = 1L, order = 1L, default = NA)

rdiffx(x, n = 1L, order = 1L, default = NA)

ldiffx(x, n = 1L, order = 1L, default = NA)
}
\description{
Returns
}
