% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill-functions.R
\name{fill_both}
\alias{fill_both}
\title{Fill with \code{locf} and \code{nocb}}
\usage{
fill_both(body, idx, order = c("locf", "nocb"))
}
\arguments{
\item{body}{\verb{[numeric vector]}

The body of the vector.}

\item{idx}{\verb{[integer vector]}

the index to replace with.}

\item{order}{\verb{[character: "locf"]}

Select which filling algorithms will occur first "locf" or "nocb".}
}
\value{
Returns a vector with the same class and attributes as the input vector.
}
\description{
Fill with \code{locf} and \code{nocb}
}
\examples{

leadx(1:4, fill = fill_both)
leadx(1:4, fill = ~ fill_both(.x,.y, order = "nocb"))

lagx(1:4, fill = fill_both)
lagx(1:4, fill = ~ fill_both(.x,.y, order = "nocb"))


}
