% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.R
\name{fill_linear}
\alias{fill_linear}
\title{Fill with linear approximation}
\usage{
fill_linear(vec, body, idx)
}
\value{
Returns a vector with the same class and attributes as the input vector.
}
\description{
Fill with linear approximation
}
