% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill-functions.R
\name{fill_locf}
\alias{fill_locf}
\title{Fill with "Last Observation Carried Forward"}
\usage{
fill_locf(body, idx, fail = NA)
}
\arguments{
\item{body}{\verb{[numeric vector]}

The body of the vector.}

\item{idx}{\verb{[integer vector]}

the index to replace with.}

\item{fail}{\verb{[numeric(1) or numeric vector: fill]}

In case it fails to fill some values.}
}
\value{
Returns a vector with the same class and attributes as the input vector.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#maturing'><img src='figures/lifecycle-maturing.svg' alt='Maturing lifecycle'></a>}}{\strong{Maturing}}
}
\examples{
x <- c(5,3,2,2,5)

lagx(x, n = 2, fill = fill_locf)

# A not so very neat way to deal with NA when `fill_locf` fails is (WIP)
lagx(x, n = 2, fill = ~ fill_locf(.x,.y, fail = 0))

leadx(x, n = 2, fill = fill_locf)

lagx(x, n = 2, fill = fill_nocb)

leadx(x, n = 2, fill = fill_nocb)

leadx(x, n = 2, fill = ~ fill_nocb(.x,.y, fail = 0))
}
