% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.R
\name{fill_spline}
\alias{fill_spline}
\title{Fill with "cubic spline interpolation"}
\usage{
fill_spline(body, idx, ...)
}
\arguments{
\item{body}{\verb{[numeric vector]}

The body of the vector.}

\item{idx}{\verb{[integer vector]}

the index to replace with.}

\item{...}{Further arguments passed to \verb{\link[stats]\{spline\}}}
}
\value{
Returns a vector with the same class and attributes as the input vector.
}
\description{
Fill with "cubic spline interpolation"
}
\examples{
x <- c(5,3,NA,2,5)
fill_spline(x, 3)
}
