% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill-functions.R
\name{fill_vec}
\alias{fill_vec}
\title{Fill with values}
\usage{
fill_vec(vec)
}
\arguments{
\item{vec}{\verb{[numeric]}

Numeric vector of the same length}
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
}
\examples{

lagx(c(1:5), fill = fill_vec(1:5))
\dontrun{

lagx(c(1:5), fill = fill_window(roll_mean(.x)))
}
}
