% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-cycle-trends.R
\name{filter_bk}
\alias{filter_bk}
\title{Baxter King Filter}
\usage{
filter_bk(
  x,
  min = NULL,
  max = NULL,
  nfix = NULL,
  type = c("fixed", "variable"),
  drift = FALSE
)
}
\arguments{
\item{x}{\verb{[univariate vector]}

Univariate vector, numeric or ts object with only one dimension.}

\item{min}{\verb{[numeric(1): NULL]}}

\item{max}{\verb{[numeric(1): NULL]}}

\item{nfix}{\verb{[logical(1): FALSE]}}
}
\value{
Returns a vector with the same class and attributes as the input vector.
}
\description{
This function computes the cyclical component of the Hodrick-Prescot filter.
}
