% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-filter.R
\name{filter_cf}
\alias{filter_cf}
\title{Christiano-Fitzgerald Filter}
\usage{
filter_cf(x, ...)
}
\arguments{
\item{x}{\verb{[univariate vector]}

Univariate vector, numeric or ts object with only one dimension.}

\item{...}{Further arguments passed to \code{\link[mFilter]{cffilter}}.}
}
\description{
This function computes the cyclical component of the Christiano-Fitzgerald filter.
}
\examples{
\donttest{
x <- cumsum(rnorm(100))
filter_cf(x)
}
}
