% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-filter.R
\name{filter_hamilton}
\alias{filter_hamilton}
\title{Hamilton Filter}
\usage{
filter_hamilton(x, n = 4, horizon = 8, fill = NA, fill_fun = NULL)
}
\arguments{
\item{x}{\verb{[univariate vector]}

Univariate vector, numeric or ts object with only one dimension.}

\item{n}{\verb{[integer(1): 4]}

A value indicating the number of lags}

\item{horizon}{\verb{[integer(1): 8]}

A value indicationg the number of periods to look ahead.}

\item{fill}{\verb{[numeric: NA]}

Value used to fill observations.}

\item{fill_fun}{\verb{[function: NULL]}

Function used to fill observations. Uses the remaining observations as the
body of the function.}

\item{...}{Additional arguments passed on to the \code{fill_fun} function.}
}
\value{
Returns a vector with the same class and attributes as the input vector.
}
\description{
This function computes the cyclical component of the Hamilton filter.
}
