% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-filter.R
\name{filter_hamilton}
\alias{filter_hamilton}
\title{Hamilton Filter}
\usage{
filter_hamilton(x, p = 4, horizon = 8, fill = NA)
}
\arguments{
\item{x}{\verb{[univariate vector]}

Univariate vector, numeric or ts object with only one dimension.}

\item{p}{\verb{[integer(1): 4]}

A value indicating the number of lags}

\item{horizon}{\verb{[integer(1): 8]}

A value indicationg the number of periods to look ahead.}

\item{fill}{\verb{[numeric or function: NA]}

Value(s) of function used to fill observations.}
}
\value{
Returns a vector with the same class and attributes as the input vector.
}
\description{
This function computes the cyclical component of the Hamilton filter.
}
\examples{
x <- cumsum(rnorm(100))
filter_hamilton(x)
}
