% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-filter.R
\name{filter_hp}
\alias{filter_hp}
\title{Hodrick-Prescot Filter}
\usage{
filter_hp(x, ...)
}
\arguments{
\item{x}{\verb{[univariate vector]}

Univariate vector, numeric or ts object with only one dimension.}

\item{...}{Further arguments passed to \code{\link[mFilter]{hpfilter}}.}
}
\description{
This function computes the cyclical component of the Hodrick-Prescot filter.
}
\examples{
\donttest{
unemp <- ggplot2::economics$unemploy
select_lambda("monthly")
unemp_cycle <- filter_hp(unemp)
plotx(unemp, unemp_cycle)
}
}
\seealso{
select_lambda
}
