% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-outliers.R
\name{out_iqr}
\alias{out_iqr}
\title{Detect outliers with Tukey's method}
\usage{
out_iqr(x, cutoff = 1.5, fill = NA, fill_fun = NULL, ...)
}
\arguments{
\item{x}{\verb{[univariate vector]}

Univariate vector, numeric or ts object with only one dimension.}

\item{cutoff}{\verb{[numeric(1): 1.5]}}

\item{fill}{\verb{[numeric: NA]}

Value used to fill observations.}

\item{fill_fun}{\verb{[function: NULL]}

Function used to fill observations. Uses the remaining observations as the
body of the function.}

\item{...}{further arguments passed to \code{quantile}.}
}
\description{
Detect outliers with Tukey's method
}
