% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-outliers.R
\name{out_iqr}
\alias{out_iqr}
\title{Detect outliers with Tukey's method}
\usage{
out_iqr(x, cutoff = 1.5, fill = NA, ...)
}
\arguments{
\item{x}{\verb{[univariate vector]}

Univariate vector, numeric or ts object with only one dimension.}

\item{cutoff}{\verb{[numeric(1): 1.5]}}

\item{fill}{\verb{[numeric or function: NA]}

Numeric value(s) or function used to fill observations.}

\item{...}{further arguments passed to \code{quantile}.}
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#maturing'><img src='figures/lifecycle-maturing.svg' alt='Maturing lifecycle'></a>}}{\strong{Maturing}}
}
\examples{

out_iqr(c(0,1,3,4,20))

}
