% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-outliers.R
\name{out_pt}
\alias{out_pt}
\title{Detect outliers with Percentiles}
\usage{
out_pt(x, pt_low = 0.1, pt_high = 0.9, fill = NA)
}
\arguments{
\item{x}{\verb{[univariate vector]}

Univariate vector, numeric or ts object with only one dimension.}

\item{pt_low}{the lowest quantile}

\item{pt_high}{the highest quantile}

\item{fill}{\verb{[numeric or function: NA]}

Numeric value(s) or function used to fill observations.}
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#maturing'><img src='figures/lifecycle-maturing.svg' alt='Maturing lifecycle'></a>}}{\strong{Maturing}}
}
\examples{
x <- c(1, 3, -1, 5, 10, 100)
out_pt(x)

}
