% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-outliers.R
\name{out_score_z}
\alias{out_score_z}
\title{Detect outliers with zscore}
\usage{
out_score_z(x, cutoff = 3, fill = NA, ...)
}
\arguments{
\item{x}{\verb{[univariate vector]}

Univariate vector, numeric or ts object with only one dimension.}

\item{cutoff}{\verb{[numeric(1): 3]}}

\item{fill}{\verb{[numeric or function: NA]}

Value(s) of function used to fill observations.}
}
\description{
Detect outliers with zscore
}
