% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-outliers.R
\name{out_score_z}
\alias{out_score_z}
\title{Detect outliers with zscore}
\usage{
out_score_z(x, cutoff = 3, fill = NA, ...)
}
\arguments{
\item{x}{\verb{[univariate vector]}

Univariate vector, numeric or ts object with only one dimension.}

\item{cutoff}{\verb{[numeric(1): 3]}}

\item{fill}{\verb{[numeric or function: NA]}

Numeric value(s) or function used to fill observations.}

\item{...}{Further arguments passed to \code{score}.}
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#maturing'><img src='figures/lifecycle-maturing.svg' alt='Maturing lifecycle'></a>}}{\strong{Maturing}}
}
\examples{
out_score_z(c(0,0.1,2,1,3,2.5,2,.5,6,4,100))

}
