% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-root-power.R
\name{pow}
\alias{pow}
\title{Exponent or the nth Power}
\usage{
pow(x, pow = NA, modulus = FALSE)
}
\arguments{
\item{x}{\verb{[univariate vector]}

Univariate vector, numeric or ts object with only one dimension.}

\item{pow}{\verb{[numeric(1): NA]}

The nth power}

\item{modulus}{positive}
}
\value{
Returns a vector with the same class and attributes as the input vector.
}
\description{
Exponent or the nth Power
}
