% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-root-power.R
\name{pow}
\alias{pow}
\title{nth Power Transformation}
\usage{
pow(x, pow = NULL, modulus = FALSE)
}
\arguments{
\item{x}{\verb{[univariate vector]}

Univariate vector, numeric or ts object with only one dimension.}

\item{pow}{\verb{[numeric(1): NA]}

The nth power.}

\item{modulus}{positive}
}
\value{
Returns a vector with the same class and attributes as the input vector.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#stable'><img src='figures/lifecycle-stable.svg' alt='Stable lifecycle'></a>}}{\strong{Stable}}
}
\examples{
pow(2, 2)
pow(-2, 2)
pow(-2,2, TRUE)
}
