% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-root-power.R
\name{pow_boxcox}
\alias{pow_boxcox}
\title{Some title}
\usage{
pow_boxcox(x, lam1, lam2 = NULL)
}
\arguments{
\item{x}{\verb{[univariate vector]}

Univariate vector, numeric or ts object with only one dimension.}

\item{lam1}{lambda1}

\item{lam2}{lambda2}
}
\value{
Returns a vector with the same class and attributes as the input vector.
}
\description{
Some title
}
\examples{
set.seed(123)
x <- rnorm(100)

}
