% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-root-power.R
\name{pow_yj}
\alias{pow_yj}
\title{Yeo and Johnson(2000)}
\usage{
pow_yj(x, lam = NULL)
}
\arguments{
\item{x}{\verb{[univariate vector]}

Univariate vector, numeric or ts object with only one dimension.}

\item{lam}{lambda}
}
\value{
Returns a vector with the same class and attributes as the input vector.
}
\description{
Yeo and Johnson(2000)
}
