% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-basic.R
\name{std}
\alias{std}
\alias{std_mean}
\alias{std_median}
\title{Standarization}
\usage{
std_mean(x, na.rm = getOption("transx.na.rm"))

std_median(x, na.rm = getOption("transx.na.rm"))
}
\arguments{
\item{x}{\verb{[univariate vector]}

Univariate vector, numeric or ts object with only one dimension.}

\item{na.rm}{\verb{[logical(1): getOption("transx.na.rm")]}

A value indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
Returns a vector with the same class and attributes as the input vector.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#maturing'><img src='figures/lifecycle-maturing.svg' alt='Maturing lifecycle'></a>}}{\strong{Maturing}}

Convert number of standard deviations by which the value of a raw score
is above or below the mean value of what is being observed or measured.
}
\examples{
x <- c(10,2,5,3)
std_mean(x)
scale(x)

std_median(x)
}
